requirements.txt
aiofiles
asyncio
aiohttp==3.11.11
beautifulsoup4
dnspython
ffmpeg-python
Flask==1.1.2
gitpython
gunicorn==20.1.0
hachoir
heroku3==5.1.4
httpx[http2]
lyricsgenius
motor
ntgcalls
pillow
psutil
py-tgcalls
pykeyboard
Kurigram==2.1.36
python-dotenv
requests
speedtest-cli
spotipy
TgCrypto
unidecode
werkzeug==2.0.2
yt-dlp
py-yt-search>=0.2
pyyaml